lint: apply reviewed unsafe fixes (SIM105, B905, F841, RUF046, RUF059, RUF022, SIM101)
Each rule applied alone with --unsafe-fixes and the diff read.
SIM105: 15 applied, no comment lost; 1 more done by hand (peer_record.py,
  ruff offers no fix there because of a trailing comment). New contextlib
  imports sorted with the safe I001 fix.
B905: 3 applied, then strict=False changed to strict=True by hand: all
  three zips are provably equal length (asserted len 5 vs a 5-literal,
  two arrays allocated and truncated together, two 3-tuples).
F841: 2 applied; the leftover bare expression F0 / k in
  test_runner_transient.py removed by hand.
RUF046: 3 applied; round() without ndigits returns int for Python floats
  and for numpy scalars (checked on numpy 2.5.3).
RUF059: 4 applied (unused unpacked names get a leading underscore).
RUF022: 0 applied, 1 reverted. Sorting core/__init__.py __all__ scatters
  its domain section comments; suppressed with a reason instead.
SIM101: 1 applied; both operands are tuples of classes.
ruff check src tests: 41 to 11. Unit 300, integration 55 unchanged.

+++ src/opensees_studio/views/docks/deformed_shape.py
@@ -87,7 +87,7 @@ class DeformedShapeView(QWidget):
     def _on_spin(self, value: float) -> None:
         if self._suggested > 0:
             multiplier = value / self._suggested
-            slider_val = max(1, min(1000, int(round(multiplier * 100))))
+            slider_val = max(1, min(1000, round(multiplier * 100)))
             self._slider.blockSignals(True)
             self._slider.setValue(slider_val)
             self._slider.blockSignals(False)
